github_sync: smarter target branch selection
Implement "sticky" target branch: if a series has an associated open
PR without merge-conflict, then do not consider other target branches.

The config allows for multiple target branches to be associated with
the same tag, for example:

  "tag_to_branch_mapping": {
    "tag": ["target1", "target2"]
  }

Whenever a series is being synced, KPD tries to apply it to each
target branch in order, stopping at the first successful apply.  Then
it opens/updates a tracking PR, and closes all other PRs for this
series.

This works fine most of the time. However, this logic may lead to
series suddenly switching branches. Example scenario:

* v1 successfully applied to target1, pr1 is created for it
  * submitter receives CI report with v1 applied to target1
* v2 fails to apply to target1, but succeeds on target2
  * KPD opens pr2 for target2, and closes pr1
  * submitter receives CI report with v2 applied to target2 with no
    indication that something happened

What should've happened in this case is pr1 labeled as
"merge-conflict" and target2 not tried.

The "stickiness" is implemented by checking existing PRs when
determining the list of target branches to apply the series to.

Signed-off-by: Ihor Solodrai <isolodrai@meta.com>