daemon: do not crash on unhandled exceptions
So far the main loop in KernelPatchesWorker.run() has been counting
how many unhandled exceptions happened concurrently. If it crossed a
threshold (hardcoded to 5), the the last caught exceptions is raised,
crashing the daemon.

This turns out to be annoying, because KPD talks to github and
patchwork via network, and so whenever there are network errors or
external outages (which happened in the past both with github and
patchwork), the KPD service falls into a crash-restart loop.

Given that KPD can not do anything if it can't reach it's
dependencies, crashes don't do anything except generating noise for
whoever is operating it.

Change the handling of unhandled (ha) exceptions:
  * record the failure and exception name in stats, which are
    periodically pushed to configured metric_logger
  * dump the stack trace of every unhandled exception to the log, so
    that they can still be investigated

Signed-off-by: Ihor Solodrai <isolodrai@meta.com>