daemon: fix metrics submission to metrics logger
In 69d4c22 the metrics submission in daemon was switched from
opentelementry to metrics logger.

The new submission is buggy however, because the metrics incremented
in KernelPatchesWorker.run() loop are immediately reset on the next
interation, before they are submitted.

Fix this by factoring out submit_metrics() and running it on every
run() loop iteration, and not only on successful sync_patches() call.

Signed-off-by: Ihor Solodrai <isolodrai@meta.com>