patchwork: fix subject name to branch name mapping
Subject.branch() returns a branch name constructed from a series id.
Fix a bug in this functions: pick the id of the last (the newest)
relevant_series() and not the first one (the oldest).

This bug was dormant most of the time, because usually there is only
one relevant series for a given subject name. However when there are
more, we got a mismatch between what patches were picked up (the
newest, correctly) and to what series branch they were applied (the
oldest, incorrectly). This caused various issues, for example
NewPRWithNoChangeException on attempt to apply the patches.

Add an integration test with a real example of a series creating this
issue. Fix other tests, that were expecting incorrect behavior.

Signed-off-by: Ihor Solodrai <ihor.solodrai@linux.dev>