Fix PR matching on cleanup
GithubSync.close_existing_prs_for_series() is supposed to close all
*other* PRs associated with a series being processed.  So far it only
looked at the *branch* name (`series/<id>=><base>`), when deciding
whether two PRs match.  This worked due to a bug, fixed in 88cb274,
which caused the oldest series id (branch name) to always be
associated with the series name (PR title).  Now that the bug is
fixed, we have to match by name during clean up, which is what happens
in this change.

Differential Revision: D79763361

Reviewed-by: Amery Hung <ameryhung@meta.com>
Signed-off-by: Ihor Solodrai <isolodrai@meta.com>