pr.update() before processing PR labels while evaluating CI results
There is an intermittent email spam issue in rc: KPD tries to update
PR labels, say from ci-fail to ci-pass, and immediately updates them
back, causing an email notification on every iteration.

The logic in the code is reasonable: we remove a "not_label" first, if
it's in current labels, and add "new_label" and send a notification if
it's not in current labels.

The only odd thing that seems to be correlated with this behavior is
forward proxy errors, which suggests failures when changing the pr
(which makes calls github API).

I noticed that KPD is looking at 2-3s old PR object in
branch_worker.evaluate_ci_result().  While generally this shouldn't be
an issue, let's add an explicit pr.update() call to make sure up to
date labels are tested.

The hypothesis is that KPD misbehavior might be caused by failed (or
implicitly retried by proxy??) pr.remove_from_labels() /
pr.add_to_labels() calls.  If this is so, then adding pr.update()
would shift the failure/retry to a less disruptive GET operation,
which will "fix" KPD even if the proxy errors are still there.

Differential Revision: D79931497

Reviewed-by: Jordan Rome <jordalgo@meta.com>
Signed-off-by: Ihor Solodrai <isolodrai@meta.com>