branch_worker: PR comments pre-processing
The format of PR comments (which is github-style markdown) may not be
suitable for plain-text email responses, depending on the use-case.

Implement a mechanism for pre-processing the contents of PR comments
before forwarding them with email.

The function transforming the content string must be a method of
BranchWorker class, and then its name can be specified (or not) in the
pr_comments_forwarding of the KPD config.

If pre-processing function is not specified, couldn't be found or
fails, the comment text remains unchanged.

Signed-off-by: Ihor Solodrai <ihor.solodrai@linux.dev>