patchwork: derive check context field from CI job name
Currently for BPF CI the "context" field of a checks is effectively a
constant with enumerator, for example:

    bpf/vmtest-bpf-next-VM_Test-48

This has an undesirable side-effect: different checks (such as "AI
review" vs "test_progs") may overwrite each other, depending on when
they run: for every PR update or for every new revision of the series.

Change how the context identifier is constructed: slugify the job
name, so that the check is associated with a job, independent of the
order or run time.

Signed-off-by: Ihor Solodrai <ihor.solodrai@linux.dev>